V0.1
New Project where the user will have to input the coordinates of the sun rotation to match the lego bricks location.

this is a simple Console app for practice, a diagram is also included.

Next Update:

Will be testing the classes in main since I'm using OOP

Bugs:

None of yet