V0.4
Uploading this before my battery goes flat, Things are coming to together using the OOP concept.

The user can also enter the location of the brick which will check for a true/false with the rotation.

However this does make most of the different obsolete so I may head back to the drawing board yet...

Bugs:

None found