V0.6
The concept is now understood in my little brain and Iv'e sadly had to taken to ChatGBT to help out the following bugs

Bugs:
Expected identifier with the return type in the bool method.

Two required ";"
(this should not be the case)

You can make an object via the child brick class but it doesn't want to call the bool method.